code/requirements.txt
absl-py==2.3.0
accelerate==1.6.0
acres==0.5.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyascii==0.3.3
anyio==4.9.0
asttokens==2.4.1
async-timeout==5.0.1
atlastk==0.13.5
attrdict==2.0.1
attrs==25.3.0
audioread==3.0.1
autopep8==2.3.2
av==14.4.0
babel==2.17.0
bangla==0.0.5
bark==0.1.5
beautifulsoup4==4.13.4
binpacking==1.5.2
bitarray==3.5.1
bitsandbytes==0.46.1
blinker==1.9.0
blis==1.2.1
blobfile==3.0.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.39.8
botocore==1.39.8
cached_path==1.7.3
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.8.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chattts==0.2.4
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
conformer==0.3.2
contourpy==1.2.1
coqpit==0.0.17
cpm-kernels==1.0.11
crcmod==1.7
cryptography==45.0.4
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dacite==1.9.2
datasets==3.3.2
dateparser==1.1.8
debugpy==1.8.7
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.9
defusedxml==0.7.1
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
easydict==1.13
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
encodec==0.1.1
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.0
# Editable install with no version control (fairseq==0.12.2)
-e your_path/LLaMA-Omni/fairseq
fastapi==0.115.13
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
fitz==0.0.1.dev2
flash_attn @ file://your_path/flash_attn-2.7.4.post1%2Bcu12torch2.6cxx11abiTRUE-cp310-cp310-linux_x86_64.whl#sha256=bfdb0f290cc3d21d0810ba49a360ef91090f62cdc1345ec6900447e0d12d99af
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
frontend==0.0.3
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2024.12.0
funcy==2.0
future==1.0.0
g2pkk==0.1.2
gdown==5.2.0
genai==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.23.1
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.73.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.4
hjson==3.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.1
humanfriendly==10.0
hydra-core==1.0.7
HyperPyYAML==1.2.2
idna==3.10
immutabledict==4.2.1
immutables==0.20
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isodate==0.7.2
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
jsonlines==1.2.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
lightning-utilities==0.14.3
# Editable install with no version control (llamafactory==0.9.4.dev0)

llvmlite==0.43.0
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mergekit==0.0.6
modelscope==1.27.1
modelscope_studio==1.2.2
more-itertools==10.7.0
mpmath==1.3.0
# Editable install with no version control (ms_swift==3.6.0.dev0)

msgpack==1.1.1
multidict==6.5.0
multiprocess==0.70.16
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==2.8.8
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
nltk==3.6.5
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.0.6
onnxruntime==1.22.1
openai==0.27.10
openai-whisper==20250625
orjson==3.10.18
oss2==2.19.1
packaging==25.0
pandas==1.5.3
parso==0.8.4
pastel==0.2.1
pathlib==1.0.1
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
poethepoet==0.10.0
pooch==1.8.2
portalocker==2.0.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
prov==2.1.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase16384==0.3.8
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.54.0
pysbd==0.3.4
PySocks==1.7.1
pytest==8.4.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.0
qwen-omni-utils==0.0.4
random_word==1.0.13
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rouge==1.0.1
rouge-chinese==1.0.3
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.0
s3transfer==0.13.0
sacrebleu==1.5.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.33.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
silero-vad==5.1.2
simplejson==3.20.1
six==1.16.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
sox==1.5.0
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
sty==1.0.6
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
textblob==0.17.1
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.3.3
timm==1.0.17
tinycss2==1.4.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchcde==0.2.5
torchdiffeq==0.2.5
torchdyn==1.0.6
torchmetrics==1.7.4
torchsde==0.2.6
torchvision==0.22.1
tornado==6.4.1
tortoise==0.1.1
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
traits==7.0.2
trampoline==0.1.2
transformers==4.52.2
transformers-stream-generator==0.0.5
triton==3.3.1
trl==0.8.6
TTS==0.22.0
typeguard==4.4.4
typer==0.16.0
typing_extensions==4.14.0
tyro==0.8.14
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.5.0
utils==1.0.2
uvicorn==0.34.3
validators==0.35.0
vector-quantize-pytorch==1.22.17
vocos==0.1.0
wandb==0.21.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
whisper==1.1.10
widgetsnbextension==4.0.14
wrapt==1.17.2
x-transformers==2.4.12
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
